+++ lib/aspera/cli/plugins/node.rb
@@ -600,7 +600,8 @@ module Aspera
               return do_bulk_operation(command: command_perm, descr: 'identifier', values: :identifier) do |one_id|
                 apifid[:api].delete("permissions/#{one_id}")
                 # notify application of deletion
-                the_app[:api].permissions_send_event(created_data: created_data, app_info: the_app, types: ['permission.deleted']) unless the_app.nil?
+                the_app = apifid[:api].app_info
+                the_app&.[](:api)&.permissions_send_event(event_data: {}, app_info: the_app, types: ['permission.deleted'])
                 {'id' => one_id}
               end
             when :create
@@ -610,11 +611,11 @@ module Aspera
               create_param['access_levels'] = Api::Node::ACCESS_LEVELS unless create_param.key?('access_levels')
               # add application specific tags (AoC)
               the_app = apifid[:api].app_info
-              the_app[:api].permissions_set_create_params(create_param: create_param, app_info: the_app) unless the_app.nil?
+              the_app&.[](:api)&.permissions_set_create_params(perm_data: create_param, app_info: the_app)
               # create permission
               created_data = apifid[:api].create('permissions', create_param)
               # notify application of creation
-              the_app[:api].permissions_send_event(created_data: created_data, app_info: the_app) unless the_app.nil?
+              the_app&.[](:api)&.permissions_send_event(event_data: created_data, app_info: the_app)
               return { type: :single_object, data: created_data}
             else Aspera.error_unreachable_line
             end